Move some methods to upload mojo
...  because it is only used there
